Create README.md
Background:
We were approached by a logistics company who suspected that they were overspending on train hires. They wanted our help to determine the optimal number of trains to hire for the upcoming quarter.

Problem:
The data were recorded on an Excel sheet, with each workbook containing a week's worth of data, resulting in 52 different workbooks with varying degrees of data inconsistency. Consolidating the data and making it consistent was a massive challenge for us.

Solution:
We turned to Python to wrangle the data and make it consistent. Once the data was clean, we fed it to Power BI to create a 2-by-2 matrix with the days of the week on the rows and the train orders on the columns. Using gradient colour coding, we calculated the average percentage utilization of each train to help the client efficiently consume the data. We also created two filters to analyse each route, one for exports and one for imports. In addition, we created visuals that showed the share of cargo by the clients, both for exports and imports.

Outcome:
Our quick turnaround time was critical to the success of the project as the client had a week to make a decision on the number of trains they needed to hire. The client appreciated our solution, which gave them the necessary insights to rationalize their train hires and save some money.